requirements.txt



# --- LangChain core & peers (linha 0.3.x) ---
langchain-core>=0.3.72,<0.4
langchain-community>=0.3,<0.4
langchain-text-splitters>=0.3.9,<0.4

# Vector store Chroma (novo pacote)
langchain-chroma>=0.1.3,<0.2
chromadb>=0.5.5,<0.6

# Embeddings HuggingFace (novo pacote)
langchain-huggingface>=0.1.0,<0.2
sentence-transformers>=3.0,<3.1
transformers>=4.41,<5
huggingface_hub>=0.23,<1

# LLM (Groq) via LangChain
langchain-groq>=0.3.0,<0.4
groq>=0.9.0,<1

# Loaders / utilitários
pymupdf>=1.24,<2           # para PyMuPDFLoader
python-dotenv>=1.0,<2

# (opcional) acelera embeddings em lotes grandes
numpy>=1.24,<3


flask


-e .
